Nicer Alert, Delete/Purge Confirmation, Progress Bar, File header comments

+++ DocumentRoot/cgi-bin/db.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python2.7
-
+#This CGI Script is for updating user notes, deleting single entries, or purging the database
 import subprocess, sys
 import json
 import cgi,cgitb
@@ -14,21 +14,25 @@ def delete_entry(id):
     conn.execute("DELETE from SCANS where ID = "+id)
     conn.commit()
     conn.close()
+    result['message'] = "Deleted entry"
 
 def purge():
     conn = sqlite3.connect(docRoot+'db/test.db')
     conn.execute("DELETE from SCANS")
     for filename in glob.glob(docRoot+'scans/*'):
         os.remove(filename)
-    os.remove(docRoot+'db.zip')
+    if(os.path.isfile(docRoot+'db.zip')): #scared about this check.. you should only purge if you've zipped....
+        os.remove(docRoot+'db.zip')
     conn.commit()
     conn.close()
+    result['message'] = "Purged database"
 
 def update(id, usernotes):
     conn = sqlite3.connect(docRoot+'db/test.db')
     conn.execute("UPDATE SCANS set USERNOTES = '"+usernotes+"' where ID = "+id)
     conn.commit()
     conn.close()
+    result['message'] = "Updated usernotes"
 
 fs = cgi.FieldStorage()
 action = fs.getvalue('action')
@@ -39,16 +43,14 @@ sys.stdout.write("\n")
 sys.stdout.write("\n")
 
 result = {}
-result['success'] = True
-result['message'] = "Called db.py"
+result['type'] = 'success'
+
 result['keys'] = ",".join(fs.keys())
 d = {}
 for k in fs.keys():
     d[k] = fs.getvalue(k)
 
 result['data'] = d
-
-
 
 if(action == 'purge'):
     purge()
